Refactor package name to model to avoid Sonar complaints
